fix: 解包后端 APIResponse wrapper + model 字段对齐
后端所有响应都包在 {success, data, error, count} 里，
App 之前直接读 response.data 拿到的是外壳而不是真正数据，
导致登录和所有 API 调用实际上都会解析失败。

- DioClient 加 onResponse 拦截器自动解包
  - success=true 时提取内层 data，count 存到 response.extra
  - success=false 时抛 DioException 走统一错误处理
- 修 dashboard_provider activeRequests 的数据访问方式
- 修 log_provider 日志列表用 response.extra 拿分页 total
- Summary/StatsResponse 的 durationSeconds 加 .toInt() 防 float 类型不匹配
- Setting model 补后端返回的 updated_at 字段

+++ lib/models/stats.dart
@@ -83,7 +83,7 @@ class StatsResponse {
     final list = (json['stats'] as List?) ?? [];
     return StatsResponse(
       stats: list.map((e) => ChannelStats.fromJson(e)).toList(),
-      durationSeconds: json['duration_seconds'] ?? 0,
+      durationSeconds: (json['duration_seconds'] ?? 0).toInt(),
       rpmStats: json['rpm_stats'] != null ? RpmStats.fromJson(json['rpm_stats']) : null,
       isToday: json['is_today'] ?? false,
     );
